screenschema/cli/requirements.txt
pyyaml>=6.0
jinja2>=3.1
jsonschema>=4.0
click>=8.0
pyserial>=3.5
watchdog>=3.0
pillow>=10.0
